requirements/production.txt
flask
flask_restful
flask_sqlalchemy
flask_migrate
flask_marshmallow
marshmallow-sqlalchemy
python-dotenv
psycopg2==2.8.6
mock